Fix primality test for very small numbers
Add special cases for primes 2, 3, 5 and 7. Proceed to Miller-Rabin and
BPSW primality tests to get the correct primality result for other small
primes that are factors in the prime products, which are used in the
initial phase of primality testing.

Previously small prime numbers up to 14951 were incorrectly reported as
composite. As this primality test has never been used for so small
numbers, there were no issues in its usage.