Reducer.h: Change abs() -> labs() to avoid implicit integer truncation
The standard abs() function defined in stdlib.h takes and returns an
'int' argument (32-bit), but we expect a 64-bit integer. When another
header file, gpu_integer.h, was also included, it provided a different
function named abs(), which was returning a 64-bit integer. Builds of
some programs such as hw_vdf_client do not include gpu_integer.h, and
this caused incorrect form reduction in very rare cases, so the problem
went unnoticed for a long time.

Change to the standard labs() function to get the correct behavior
regardless of extra include files.